Render helper refactor
