+++ app/client/src/widgets/TableWidgetV2/component/renderHelpers/HeaderCellRenderer.tsx
@@ -1,0 +1,89 @@
+import React from "react";
+import { AnyStyledComponent } from "styled-components";
+
+import { DraggableHeaderWrapper } from "../TableStyledWrappers";
+import { Colors } from "constants/Colors";
+import styled from "constants/DefaultTheme";
+import { ControlIcons } from "icons/ControlIcons";
+
+const AscendingIcon = styled(ControlIcons.SORT_CONTROL as AnyStyledComponent)`
+  padding: 0;
+  position: relative;
+  top: 3px;
+  cursor: pointer;
+  transform: rotate(180deg);
+  && svg {
+    path {
+      fill: ${Colors.LIGHT_GREYISH_BLUE};
+    }
+  }
+`;
+
+const DescendingIcon = styled(ControlIcons.SORT_CONTROL as AnyStyledComponent)`
+  padding: 0;
+  position: relative;
+  top: 3px;
+  cursor: pointer;
+  && svg {
+    path {
+      fill: ${Colors.LIGHT_GREYISH_BLUE};
+    }
+  }
+`;
+
+export function TableHeaderCell(props: {
+  columnName: string;
+  columnIndex: number;
+  isHidden: boolean;
+  isAscOrder?: boolean;
+  sortTableColumn: (columnIndex: number, asc: boolean) => void;
+  isResizingColumn: boolean;
+  column: any;
+  editMode?: boolean;
+  isSortable?: boolean;
+}) {
+  const { column, editMode, isSortable } = props;
+  const handleSortColumn = () => {
+    if (props.isResizingColumn) return;
+    let columnIndex = props.columnIndex;
+    if (props.isAscOrder === true) {
+      columnIndex = -1;
+    }
+    const sortOrder =
+      props.isAscOrder === undefined ? false : !props.isAscOrder;
+    props.sortTableColumn(columnIndex, sortOrder);
+  };
+  const disableSort = editMode === false && isSortable === false;
+
+  return (
+    <div
+      {...column.getHeaderProps()}
+      className="th header-reorder"
+      onClick={!disableSort && props ? handleSortColumn : undefined}
+    >
+      <DraggableHeaderWrapper
+        className={!props.isHidden ? `draggable-header` : "hidden-header"}
+        horizontalAlignment={column.columnProperties.horizontalAlignment}
+      >
+        {props.columnName}
+      </DraggableHeaderWrapper>
+      {props.isAscOrder !== undefined ? (
+        <div>
+          {props.isAscOrder ? (
+            <AscendingIcon height={16} width={16} />
+          ) : (
+            <DescendingIcon height={16} width={16} />
+          )}
+        </div>
+      ) : null}
+      <div
+        {...column.getResizerProps()}
+        className={`resizer ${column.isResizing ? "isResizing" : ""}`}
+        onClick={(e: React.MouseEvent<HTMLElement>) => {
+          e.preventDefault();
+          e.stopPropagation();
+        }}
+      />
+    </div>
+  );
+}
